market-bot2-python-client/requirements.txt
ethereal-sdk
python-dotenv
flask
flask-cors